backend/requirements.txt
django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
cryptography>=41.0
django-cors-headers>=4.3
